Include Docker support, fix some bugs, add columns, refactor some code
I've added Dockerfile, fixed a bug (check for !destination on DragEnd), added multiple columns, refactored the DragEnd code a bit for a switch case (easier for multiple columns).
